Broken Down Style Guide Page into Components
